Grammar is changed.
I realized that it would be very difficult to define accurate grammar
based on the stile I used. So, I saw few real grammars and decide to
rewrite Expression, SimpleExpression, Assignment, Relation and Equality
Expressions in the way that it would look like real grammar. Few things
are changed in TranslationVisitor class, like changing the names of context
nodes.